pyproject.toml
[project]
authors = [
    { name = "The Board of Trustees of the Leland Stanford Junior University, through SLAC National Accelerator Laboratory"},
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "h5py",
    "numpy",
    "pydantic",
    "pyyaml",
    "pyepics",
    "slac-db @ git+ssh://git@github.com/slaclab/slac-db.git"
]
description = "Tools to support high level application development at LCLS using Python"
dynamic = ["version"]
name = "slac-devices"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"